feat(settings): add storage filepath settings
This patch adds a new storage settings object with a single field to
control the storage filepath.

One use case is to set the filepath to the empy string, which will
disable reading and writing of the servers.json file. This is useful for
deployments that want to minimize I/O and want to set tight memory
limits (writing the servers.json spikes memory).

Fixes #2074